Move make_moons generator into shared bench utilities
The DBSCAN benchmark's make_moons data generator (two interleaving
half circles, after sklearn) is also useful for other clustering
benchmarks, so move it to geo-benches/src/utils/random.rs and import
it from there.
